backend/requirements.txt
fastapi
uvicorn
openai
PyMuPDF
sentence-transformers
faiss-cpu
numpy
PyPDF2